requirements.txt
pandas==2.2.1
yfinance==0.2.37
alpaca-py==0.19.1
pandas-ta==0.3.14b0
scikit-learn==1.4.1.post1
numpy==1.26.4
